Update games-content-spanish.json
I made some modifications to the charades array. A few typos and some other translations.  For Example:
"Home Alone" is known for "Mi Pobre Angelito" and not the literal translation of "Solo en Casa"
As far as I'm concerned the TV series called "Friends" is also refered to "Friends" on spanish TV 
Even though Metastasis is based on Breaking Bad it is a different series because the cast was different.